+++ NFlags/Arguments/PrefixedDefaultValueArgument.cs
@@ -1,17 +1,17 @@
 namespace NFlags.Arguments
 {
     /// <summary>
-    /// Base class for arguments with abreviation
+    /// Base class for arguments with abbreviation
     /// </summary>
     public abstract class PrefixedDefaultValueArgument : DefaultValueArgument
     {
         /// <summary>
-        /// Abreviation name for help printing and shorthand usage.
+        /// Abbreviation name for help printing and shorthand usage.
         /// </summary>
         public string Abr;
 
         /// <summary>
-        /// Defines if flag persist for subcommands
+        /// Defines if flag persist for sub commands
         /// </summary>
         public bool IsPersistent;
     }